+++ src/routes/(protected)/(navbar)/settings/SettingsNavBar.svelte
@@ -37,6 +37,8 @@
 			Delete Account
 		{:else if page.route.id === "/(protected)/(navbar)/settings/(subpage)/app"}
 			App Settings
+		{:else if page.route.id === "/(protected)/(navbar)/settings/(subpage)/app/notifications"}
+			Notifications
 		{:else if page.route.id === "/(protected)/(navbar)/settings/(subpage)/stats"}
 			Stats
 		{/if}